@override_review Udemy_Exercise 1 - Minor changes

+++ src/Udemy/LearnProgramming/PoundsToKilograms.java
@@ -6,8 +6,8 @@ public class PoundsToKilograms {
     public static void main(String[] args){
         Scanner input = new Scanner(System.in);
         System.out.print("Enter the number of pounds:");
-        int Pounds = input.nextInt();
-        double Kilograms = Pounds * 0.45359237;
+        double Pounds = input.nextDouble();
+        double Kilograms = Pounds * 0.45359237d;
         System.out.println(Pounds+" pound is equal to "+ Kilograms +" kilograms");
     }
 }
